Fixes glibc abort when exceeding 1024 FDs.
FD_SET() was triggering an abort, when built with a newer glibc.  I
dropped in sys/select.h and bits/select.h from glibc-2.14.1, which do
not provide these checks.

Although we override __FD_SETSIZE with a custom typesizes.h, glibc was
compiled with __FD_SETSIZE = 1024, and it was aborting based on that
check.